rrtmgp_tests now read abs coef data using netcdf-cxx
The rrtmgp_tests are now capable of reading the absorption coefficient data that is shipped with RRTMGP. Currently, this uses the custom "simple" netCDF IO interface included with YAKL, which takes YAKL arrays as arguments and populates them with data read in using the netCDF-cxx interface. This seems to be the easiest way to handle this case, since netcdf-cxx provides for arbitrary data types being passed to the netcdf routines, but requires netcdf-cxx be built on the machine SCREAM is run on. This is currently not standard practice, and we may want to get away from this dependency and directly use either netcdf-c or SCORPIO. SCORPIO would be overkill here, since this is a read-once type of thing. In any event, once netcdf-cxx is available, the tests can be configured and build by passing the ``-DCMAKE_CXX_FLAGS="`nc-config --cflags` `ncxx4-config --cflags`"`` and ``-DCMAKE_EXE_LINKER_FLAGS="`nc-config --libs` `ncxx4-config --libs`"`` flags to cmake.