Make GasOpticsRRTMGP objects persist
Make the k_dist_sw and k_dist_lw instances of the GasOpticsRRTMGP class
persist so that we can only load the absorption coefficient data once
during a run. This is handled in `scream::rrtmgp::initialize()`, which
will populate the instances of k_dist_sw and k_dist_lw that live in the
`scream::rrtmgp` namespace. For some reason, I could not get to this to
work without changing the C++ standard version for the entire project.
Adding the includes for the RRTMGP headers to the scream_rrtmgp_interface
files somehow caused some standards conflicts with included yakl headers,
which should have gotten the -std=c++14 flag. I tried a number of different
CMake implementations but ultimately the only thing I could get to work was
to just set the `CMAKE_CXX_STANDARD` to 14 in the top-level CMakeLists.txt.
Also, I believe there is a bug in the RRTMGP mo_gas_optics_rrtmgp.h file,
that misses including mo_gas_concentrations.h.